Implement DBConnector::publish(), DBInterface::_unavailable_data_handler()
